update installation is async
