+++ AnnoDesigner/App.xaml
@@ -1,6 +1,7 @@
 ﻿<Application x:Class="AnnoDesigner.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             Startup="Application_Startup">
     <Application.Resources>
 
     </Application.Resources>
