+++ AnnoDesigner/UpdateWindow.xaml
@@ -1,0 +1,25 @@
+﻿<Window x:Class="AnnoDesigner.UpdateWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:AnnoDesigner"
+        xmlns:xctk="http://schemas.xceed.com/wpf/xaml/toolkit"
+        mc:Ignorable="d"
+        Title="Anno Designer"
+        Icon="/AnnoDesigner;component/icon.ico"
+        WindowStartupLocation="CenterScreen"
+        Height="200"
+        Width="250"
+        WindowStyle="None"
+        ResizeMode="NoResize">
+    <Window.Background>
+        <ImageBrush ImageSource="/AnnoDesigner;component/icon.ico" />
+    </Window.Background>
+    <Grid>
+        <xctk:BusyIndicator x:Name="busyIndicator"
+                            IsBusy="True"
+                            BusyContent="Installing Update ...">
+        </xctk:BusyIndicator>
+    </Grid>
+</Window>
